功能測試，微調bug
- 沒登入不能登出
- 修改密碼的欄位顯示改成圖形，加sessionStorage.clear()
- 信件紀錄的_blank
- 註冊的bug測試
- 寄信加一個script連結

+++ project-front/code1/components/layout/MainNavigation.js
@@ -42,14 +42,18 @@ function MainNavigation() {
               <a style={{textDecoration: 'none'}}
                 onClick={e => {
                   e.preventDefault()
-                  console.log('Logout');
+                  if(sessionStorage.getItem("token") == null){
+                    alert('尚未登入，不能登出')
+                  }else{
+                    console.log('Logout');
 
-                  // CLEAR DATA FROM STORAGE
-                  sessionStorage.clear();
-                  // sessionStorage.removeItem('token');
+                    // CLEAR DATA FROM STORAGE
+                    sessionStorage.clear();
+                    // sessionStorage.removeItem('token');
 
-                  alert("登出成功");
-                  router.push("/");
+                    alert("登出成功");
+                    router.push("/");
+                  }
                 }}>
                 登出
               </a>
